Move shared code to internet testing.
